+++ test/test.js
@@ -54,15 +54,12 @@ describe('SvgSprite', function () {
             instance.register('src/icons');
 
             assert.deepStrictEqual(
-                instance.config,
+                instance.config[0],
                 {
                     path: path.resolve(__dirname, 'src/icons'),
                     loaderOptions: {
                         spriteFilename: path.join('images', 'sprite.svg'),
                         foo: 'bar'
-                    },
-                    pluginOptions: {
-                        foo: 'baz'
                     },
                 }
             );
@@ -70,18 +67,41 @@ describe('SvgSprite', function () {
 
         it('uses the given options', function () {
             let instance = new SvgSprite();
-            instance.register('src/icons', 'dist/icon-sprite.svg', {custom: true}, {lorem: 'ipsum'});
+            instance.register('src/icons', 'dist/icon-sprite.svg', {custom: true});
 
             assert.deepStrictEqual(
-                instance.config,
+                instance.config[0],
                 {
                     path: path.resolve(__dirname, 'src/icons'),
                     loaderOptions: {
                         spriteFilename: 'dist/icon-sprite.svg',
                         custom: true
                     },
-                    pluginOptions: {
-                        lorem: 'ipsum'
+                }
+            );
+        });
+
+        it('can register multiple sprites', function () {
+            let instance = new SvgSprite();
+            instance.register('src/foo', 'dist/foo-sprite.svg', {});
+            instance.register('src/bar', 'dist/bar-sprite.svg', {});
+
+            assert.strictEqual(instance.config.length, 2);
+            assert.deepStrictEqual(
+                instance.config[0],
+                {
+                    path: path.resolve(__dirname, 'src/foo'),
+                    loaderOptions: {
+                        spriteFilename: 'dist/foo-sprite.svg',
+                    },
+                }
+            );
+            assert.deepStrictEqual(
+                instance.config[1],
+                {
+                    path: path.resolve(__dirname, 'src/bar'),
+                    loaderOptions: {
+                        spriteFilename: 'dist/bar-sprite.svg',
                     },
                 }
             );
@@ -174,6 +194,66 @@ describe('SvgSprite', function () {
             instance.register('src/icons');
             instance.webpackConfig(webpackConfig);
 
+            assert.deepStrictEqual(
+                webpackConfig.module.rules.pop(),
+                {
+                    include: [
+                        path.resolve(__dirname, 'src/icons')
+                    ],
+                    test: /\.(svg)(\?.*)?$/,
+                    use: [
+                        {
+                            loader: 'svg-sprite-loader',
+                            options: {
+                                foo: 'bar',
+                                spriteFilename: path.join('images', 'sprite.svg')
+                            }
+                        },
+                        {
+                            loader: 'svgo-loader',
+                            options: {}
+                        }
+                    ]
+                }
+            );
+        });
+
+        it('adds multiple rules for multiple sprites', function () {
+            let webpackConfig = {
+                module: {
+                    rules: []
+                }
+            };
+
+            let instance = new SvgSprite();
+            instance.register('src/icons');
+            instance.register('src/other-icons', 'images/other-sprite.svg');
+            instance.webpackConfig(webpackConfig);
+
+            assert.strictEqual(webpackConfig.module.rules.length, 2);
+            // N.B. assertion in reverse order
+            assert.deepStrictEqual(
+                webpackConfig.module.rules.pop(),
+                {
+                    include: [
+                        path.resolve(__dirname, 'src/other-icons')
+                    ],
+                    test: /\.(svg)(\?.*)?$/,
+                    use: [
+                        {
+                            loader: 'svg-sprite-loader',
+                            options: {
+                                foo: 'bar',
+                                spriteFilename: 'images/other-sprite.svg'
+                            }
+                        },
+                        {
+                            loader: 'svgo-loader',
+                            options: {}
+                        }
+                    ]
+                }
+            );
             assert.deepStrictEqual(
                 webpackConfig.module.rules.pop(),
                 {